Refactor browser_stack_utils.rb
