+++ bin/download-logs
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 require_relative '../lib/maze'
-require_relative '../lib/maze/browser_stack_utils'
+require_relative '../lib/maze/farm/browser_stack/utils'
 require_relative '../lib/maze/logger'
 require 'net/http'
 require 'json'
@@ -44,9 +44,9 @@ class DownloadLogEntry
     end
 
 
-    build_info = Maze::BrowserStackUtils.build_info username,
-                                                    access_key,
-                                                    opts[:build_id]
+    build_info = Maze::Farm::BrowserStack::Utils.build_info username,
+                                                            access_key,
+                                                            opts[:build_id]
 
     $logger.info "Getting logs for build: #{opts[:build_id]}"
 
@@ -55,20 +55,18 @@ class DownloadLogEntry
 
       log_name = "#{opts[:build_id]}-#{index + 1}"
 
-      Maze::BrowserStackUtils.download_log username,
-                                           access_key,
-                                           log_name,
-                                           session['automation_session']['device_logs_url'],
-                                           :device
+      Maze::Farm::BrowserStack::Utils.download_log username,
+                                                   access_key,
+                                                   log_name,
+                                                   session['automation_session']['device_logs_url'],
+                                                   :device
 
       $logger.info "Downloading Appium Logs for Session: #{index + 1}"
-      Maze::BrowserStackUtils.download_log username,
-                                           access_key,
-                                           log_name,
-                                           session['automation_session']['appium_logs_url'],
-                                           :appium
-
-
+      Maze::Farm::BrowserStack::Utils.download_log username,
+                                                   access_key,
+                                                   log_name,
+                                                   session['automation_session']['appium_logs_url'],
+                                                   :appium
     end
   end
 end
