+++ bin/maze-runner
@@ -10,7 +10,6 @@ require_relative '../lib/maze'
 require_relative '../lib/maze/appium_server'
 require_relative '../lib/maze/bitbar_utils'
 require_relative '../lib/maze/bitbar_devices'
-require_relative '../lib/maze/browser_stack_utils'
 require_relative '../lib/maze/bugsnag_config'
 require_relative '../lib/maze/capabilities'
 require_relative '../lib/maze/compare'
@@ -18,6 +17,7 @@ require_relative '../lib/maze/docker'
 require_relative '../lib/maze/document_server'
 require_relative '../lib/maze/errors'
 require_relative '../lib/maze/farm/browser_stack/appium/devices'
+require_relative '../lib/maze/farm/browser_stack/utils'
 require_relative '../lib/maze/helper'
 require_relative '../lib/maze/logger'
 require_relative '../lib/maze/macos_utils'
